Ignore consumer_offsets

+++ pkg/cmd/recreate.go
@@ -32,6 +32,9 @@ func (g *Recreate) Run(globals *Globals) error {
 	// Delete topics on target cluster
 	for name, _ := range topics {
 		if !g.DryRun {
+			if name == "__consumer_offsets" {
+				continue
+			}
 			err := admin.DeleteTopic(name)
 			if err != nil {
 				fmt.Printf("Unable to delete topic %s with err %s\n", name, err.Error())
@@ -42,6 +45,9 @@ func (g *Recreate) Run(globals *Globals) error {
 	}
 	// Create topics on target cluster
 	for name, topic := range topics {
+		if name == "__consumer_offsets" {
+			continue
+		}
 		configEntries := map[string]*string{}
 		for k, v := range topic.ConfigEntries {
 			switch k {
